Get ACE, core and coupled tests passing on MPS (#1672)
I fixed/skipped all relevant tests in `fme/core`, `fme/ace`, and
`fme/coupled` when calling `FME_USE_MPS=1 pytest`.
I didn't touch `fme/downscaling` since there is a lot in flux there
right now.

+++ fme/ace/aggregator/inference/annual.py
@@ -351,7 +351,10 @@ def to_dataset(data: TensorMapping, time: xr.DataArray) -> xr.Dataset:
     data_vars = {}
     for name, tensor in data.items():
         data_vars[name] = (["sample", "time"], tensor)
-    data_vars["counts"] = (["sample", "time"], np.ones(shape=time.shape))
+    data_vars["counts"] = (
+        ["sample", "time"],
+        np.ones(shape=time.shape, dtype=np.float32),
+    )
     return xr.Dataset(data_vars, coords={"valid_time": time})
